requirements.txt
future
requests==2.22
tqdm
bs4
yapic.json